requirements.txt
Flask
Flask-WTF
pymongo
Flask-Login
email-validator
paho-mqtt
qrcode
pillow
python-dotenv
mongomock
pytest
